chore: moves from utilities to multipleTemplate file and creates hasRewritesConfigForSlug logic
